Write introduction to Finch hardware
Wrote an introduction to the Finch hardware.  This describes the Finch's
sensors and actuators in a way that's more kid-friendly than the
official finch robot webpage.  I borrowed the hardware diagram from the
official site.

Also set up a table of contents which can be used to reach the
introduction page.