Document setting up Scratch with Finch
Added a part on "Programming Finch with Scratch".  Documented how to get
the BirdBrain Robot Server working, how to install Scratch, and how to
import the Finch Scratch blocks.  This is essentially a more concise
version of the finchrobot.com instructions that focuses on the CoderDojo
environment.

Also added an explanation of how BirdBrain Robot Server works since
that's not clear from the finchrobot.com site.